Fixes for CSRF and minor refactoring
