+++ .env
@@ -8,7 +8,7 @@ PUBLIC_URL = https://dev.validate.buildingsmart.org
 MEDIA_ROOT = /files_storage
 DJANGO_DB = postgresql
 DJANGO_ALLOWED_HOSTS = dev.validate.buildingsmart.org
-DJANGO_CSRF_TRUSTED_ORIGINS = https://dev.validate.buildingsmart.org https://authentication.buildingsmart.org
+DJANGO_TRUSTED_ORIGINS = https://dev.validate.buildingsmart.org https://authentication.buildingsmart.org
 DJANGO_LOG_LEVEL = INFO
 
 # DB
