+++ backend/.env
@@ -16,7 +16,7 @@ PUBLIC_URL = http://localhost:3000
 MEDIA_ROOT = .dev/files_storage
 DJANGO_DB = postgresql
 TEST_DJANGO_DB = sqlite
-DJANGO_CSRF_TRUSTED_ORIGINS = http://localhost:3000 http://localhost
+DJANGO_TRUSTED_ORIGINS = http://localhost:3000 http://localhost http://localhost:8000
 DJANGO_LOG_FOLDER = .dev/logging
 DJANGO_LOG_LEVEL = INFO
