+++ docker-compose.full.yml
@@ -40,7 +40,7 @@ services:
             CELERY_RESULT_BACKEND_DB: "db+postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@db/${POSTGRES_NAME}"
             DJANGO_DB: ${DJANGO_DB}
             DJANGO_ALLOWED_HOSTS: ${DJANGO_ALLOWED_HOSTS}
-            DJANGO_CSRF_TRUSTED_ORIGINS: ${DJANGO_CSRF_TRUSTED_ORIGINS}
+            DJANGO_TRUSTED_ORIGINS: ${DJANGO_TRUSTED_ORIGINS}
             POSTGRES_HOST: db
             POSTGRES_PORT: ${POSTGRES_PORT}
             POSTGRES_NAME: ${POSTGRES_NAME}
